fix: added cast to int back

+++ backend/metrics.py
@@ -40,15 +40,15 @@ class Metrics(object):
     
     def getTotalItemsProduced(self) -> int:
         return self.totalItemsProduced
-    #Removed cast to int
-    def getCoolantLevelPercent(self) -> float:
-        return self.coolantLevelPercent
-    #Removed cast to int
-    def getPowerConsumptionKWH(self) -> float:
-        return self.powerConsumptionKWH
-    #Removed cast to int
-    def getLaserModulePowerWeardown(self) -> float:
-        return self.laserModuleWeardownPercent
+
+    def getCoolantLevelPercent(self) -> int:
+        return (int) (self.coolantLevelPercent)
+
+    def getPowerConsumptionKWH(self) -> int:
+        return (int)(self.powerConsumptionKWH)
+
+    def getLaserModulePowerWeardown(self) -> int:
+        return (int)(self.laserModuleWeardownPercent)
     
     def getTimePerItem(self) -> int:
         return self.timePerItem
